# Conflicts:
#	AnytimeGear/AnytimeGear.Server/Models/Product.cs
#	AnytimeGear/AnytimeGear.Server/Models/Rental.cs
#	AnytimeGear/AnytimeGear.Server/Program.cs
#	AnytimeGear/anytimegear.client/src/components/Shared/NavigationBar.tsx
#	AnytimeGear/anytimegear.client/src/utils/router.tsx

+++ AnytimeGear/AnytimeGear.Server/Controllers/RentalsController.cs
@@ -33,7 +33,7 @@ public class RentalsController : ApiController
             {
                 Id = rental.Id,
                 ProductName = rental.Product.Name,
-                Price = (int)(rental.Product.Price * rental.Quantity * (rental.EndPeriod - rental.StartPeriod).TotalDays),
+                Price = (int)(rental.Product.Price * rental.Quantity * (int)(rental.EndPeriod - rental.StartPeriod).TotalDays),
                 StartDate = rental.StartPeriod.ToString("dd-MM-yyyy"),
                 EndDate = rental.EndPeriod.ToString("dd-MM-yyyy"),
                 Quantity = rental.Quantity,
